Moved UserDefinedAction parameter resolving into atomicbehaviours to make ParameterAction work
