+++ srunner/tools/openscenario_parser.py
@@ -72,6 +72,7 @@ from srunner.scenariomanager.scenarioatomics.atomic_trigger_conditions import (I
 from srunner.scenariomanager.timer import TimeOut, SimulationTimeCondition
 from srunner.tools.py_trees_port import oneshot_behavior
 from srunner.tools.scenario_helper import get_offset_transform, get_troad_from_transform
+from srunner.tools.scenario_param_ref import ParameterRef
 
 
 def oneshot_with_check(variable_name, behaviour, name=None):
@@ -87,165 +88,7 @@ def oneshot_with_check(variable_name, behaviour, name=None):
     return oneshot_behavior(variable_name, behaviour, name)
 
 
-class ParameterRef:
-    """
-    This class stores osc parameter reference in its original form.
-    Returns the converted value whenever it is used.
-    """
 
-    def __init__(self, reference_text) -> None:
-        # TODO: (for OSC1.1) add methods(lexer and math_interpreter) to
-        #  recognize and interpret math expression from reference_text
-        self.reference_text = str(reference_text)
-
-    def is_literal(self) -> bool:
-        """
-        Returns: True when text is a literal/number
-        """
-        return self._is_matching(pattern=r"(-)?\d+(\.\d*)?")
-
-    def is_parameter(self) -> bool:
-        """
-        Returns: True when text is a parameter
-        """
-        return self._is_matching(pattern=r"[$][A-Za-z_][\w]*")
-
-    def _is_matching(self, pattern: str) -> bool:
-        """
-        Returns: True when pattern is matching with text
-        """
-        match = re.search(pattern, self.reference_text)
-        if match is not None:
-            matching_string = match.group()
-            return matching_string == self.reference_text
-        return False
-
-    def _cast_type(self, other):
-        try:
-            if isinstance(other, str):
-                return self.__str__()
-            elif isinstance(other, float):
-                return self.__float__()
-            elif isinstance(other, int):
-                return self.__int__()
-            else:
-                raise Exception("Type conversion for type {} not implemented".format(type(other)))
-        except ValueError:
-            # If the parameter value can't be converted into type of 'other'
-            return None
-
-    def get_interpreted_value(self):
-        """
-        Returns: interpreted value from reference_text
-        """
-        if self.is_literal():
-            value = self.reference_text
-        elif self.is_parameter():
-            value = CarlaDataProvider.get_osc_global_param_value(self.reference_text)
-            if value is None:
-                raise Exception("Parameter '{}' is not defined".format(self.reference_text[1:]))
-        else:
-            value = None
-        return value
-
-    def __float__(self) -> float:
-        value = self.get_interpreted_value()
-        if value is not None:
-            return float(value)
-        else:
-            raise Exception("could not convert '{}' to float".format(self.reference_text))
-
-    def __int__(self) -> int:
-        value = self.get_interpreted_value()
-        if value is not None:
-            return int(float(value))
-        else:
-            raise Exception("could not convert '{}' to int".format(self.reference_text))
-
-    def __str__(self) -> str:
-        value = self.get_interpreted_value()
-        return str(value) if value is not None else self.reference_text
-
-    def __repr__(self):
-        value = self.get_interpreted_value()
-        return value if value is not None else self.reference_text
-
-    def __hash__(self):
-        value = self.get_interpreted_value()
-        return hash(value) if value is not None else hash(self.reference_text)
-
-    def __radd__(self, other) -> bool:
-        return other + self.__float__()
-
-    def __add__(self, other) -> bool:
-        return other + self.__float__()
-
-    def __rsub__(self, other) -> bool:
-        return other - self.__float__()
-
-    def __sub__(self, other) -> bool:
-        return self.__float__() - other
-
-    def __rmul__(self, other) -> bool:
-        return other * self.__float__()
-
-    def __mul__(self, other) -> bool:
-        return other * self.__float__()
-
-    def __truediv__(self, other) -> bool:
-        return self.__float__() / other
-
-    def __rtruediv__(self, other) -> bool:
-        return other / self.__float__()
-
-    def __eq__(self, other) -> bool:
-        value = self._cast_type(other)
-        return other == value
-
-    def __ne__(self, other) -> bool:
-        value = self._cast_type(other)
-        return other != value
-
-    def __ge__(self, other) -> bool:
-        value = self._cast_type(other)
-        return value >= other
-
-    def __le__(self, other) -> bool:
-        value = self._cast_type(other)
-        return value <= other
-
-    def __gt__(self, other) -> bool:
-        value = self._cast_type(other)
-        return value > other
-
-    def __lt__(self, other) -> bool:
-        value = self._cast_type(other)
-        return value < other
-
-    def __GE__(self, other) -> bool:  # pylint: disable=invalid-name
-        value = self._cast_type(other)
-        return value >= other
-
-    def __LE__(self, other) -> bool:  # pylint: disable=invalid-name
-        value = self._cast_type(other)
-        return value <= other
-
-    def __GT__(self, other) -> bool:  # pylint: disable=invalid-name
-        value = self._cast_type(other)
-        return value > other
-
-    def __LT__(self, other) -> bool:  # pylint: disable=invalid-name
-        value = self._cast_type(other)
-        return value < other
-
-    def __iadd__(self, other) -> bool:
-        return self.__float__() + other
-
-    def __isub__(self, other) -> bool:
-        return self.__float__() - other
-
-    def __abs__(self):
-        return abs(self.__float__())
 
 
 class OpenScenarioParser(object):
@@ -389,7 +232,9 @@ class OpenScenarioParser(object):
         returns:
             Catalog entry (XML ElementTree)
         """
-        entry = catalogs[ParameterRef(catalog_reference.attrib.get("catalogName"))][ParameterRef(catalog_reference.attrib.get("entryName"))]
+        catalog_name = ParameterRef(catalog_reference.attrib.get("catalogName"))
+        catalog_entry = ParameterRef(catalog_reference.attrib.get("entryName"))
+        entry = catalogs[catalog_name][catalog_entry]
         entry_copy = copy.deepcopy(entry)
         catalog_copy = copy.deepcopy(catalog_reference)
         entry = OpenScenarioParser.assign_catalog_parameters(entry_copy, catalog_copy)
@@ -664,7 +509,7 @@ class OpenScenarioParser(object):
             droll = 0
             if rel_pos.find('Orientation') is not None:
                 orientation = rel_pos.find('Orientation')
-                is_absolute = (orientation.attrib.get('type') == "absolute")
+                is_absolute = (str(ParameterRef(orientation.attrib.get('type'))) == "absolute")
                 dyaw = math.degrees(float(ParameterRef(orientation.attrib.get('h', 0))))
                 dpitch = math.degrees(float(ParameterRef(orientation.attrib.get('p', 0))))
                 droll = math.degrees(float(ParameterRef(orientation.attrib.get('r', 0))))
@@ -833,23 +678,22 @@ class OpenScenarioParser(object):
         else:
             raise AttributeError("Unknown position")
 
+    
     @staticmethod
     def convert_condition_to_atomic(condition, actor_list):
         """
         Convert an OpenSCENARIO condition into a Behavior/Criterion atomic
-
         If there is a delay defined in the condition, then the condition is checked after the delay time
         passed by, e.g. <Condition name="" delay="5">.
-
         Note: Not all conditions are currently supported.
         """
 
         atomic = None
         delay_atomic = None
-        condition_name = condition.attrib.get('name')
+        condition_name = str(ParameterRef(condition.attrib.get('name')))
 
-        if condition.attrib.get('delay') is not None and float(condition.attrib.get('delay')) != 0:
-            delay = float(condition.attrib.get('delay'))
+        if condition.attrib.get('delay') is not None and float(ParameterRef(condition.attrib.get('delay'))) != 0:
+            delay = float(ParameterRef(condition.attrib.get('delay')))
             delay_atomic = TimeOut(delay)
 
         if condition.find('ByEntityCondition') is not None:
@@ -860,14 +704,15 @@ class OpenScenarioParser(object):
             for triggering_entities in condition.find('ByEntityCondition').iter('TriggeringEntities'):
                 for entity in triggering_entities.iter('EntityRef'):
                     for actor in actor_list:
-                        if actor is not None and entity.attrib.get('entityRef', None) == actor.attributes['role_name']:
+                        ref_name = str(ParameterRef(entity.attrib.get('entityRef', None)))
+                        if (actor is not None and ref_name == actor.attributes['role_name']):
                             trigger_actor = actor
                             break
 
             for entity_condition in condition.find('ByEntityCondition').iter('EntityCondition'):
                 if entity_condition.find('EndOfRoadCondition') is not None:
                     end_road_condition = entity_condition.find('EndOfRoadCondition')
-                    condition_duration = ParameterRef(end_road_condition.attrib.get('duration'))
+                    condition_duration = float(ParameterRef(end_road_condition.attrib.get('duration')))
                     atomic_cls = py_trees.meta.inverter(EndofRoadTest)
                     atomic = atomic_cls(trigger_actor, condition_duration, terminate_on_failure=True,
                                         name=condition_name)
@@ -879,20 +724,21 @@ class OpenScenarioParser(object):
                         collision_entity = collision_condition.find('EntityRef')
 
                         for actor in actor_list:
-                            if collision_entity.attrib.get('entityRef', None) == actor.attributes['role_name']:
+                            if (str(ParameterRef(collision_entity.attrib.get('entityRef', None)))
+                                    == actor.attributes['role_name']):
                                 triggered_actor = actor
                                 break
 
                         if triggered_actor is None:
                             raise AttributeError("Cannot find actor '{}' for condition".format(
-                                collision_condition.attrib.get('entityRef', None)))
+                                str(ParameterRef(collision_condition.attrib.get('entityRef', None)))))
 
                         atomic_cls = py_trees.meta.inverter(CollisionTest)
                         atomic = atomic_cls(trigger_actor, other_actor=triggered_actor, terminate_on_failure=True,
                                             name=condition_name)
 
                     elif collision_condition.find('ByType') is not None:
-                        collision_type = collision_condition.find('ByType').attrib.get('type', None)
+                        collision_type = str(ParameterRef(collision_condition.find('ByType').attrib.get('type', None)))
 
                         triggered_type = OpenScenarioParser.actor_types[collision_type]
 
@@ -906,7 +752,7 @@ class OpenScenarioParser(object):
 
                 elif entity_condition.find('OffroadCondition') is not None:
                     off_condition = entity_condition.find('OffroadCondition')
-                    condition_duration = ParameterRef(off_condition.attrib.get('duration'))
+                    condition_duration = float(ParameterRef(off_condition.attrib.get('duration')))
                     atomic_cls = py_trees.meta.inverter(OffRoadTest)
                     atomic = atomic_cls(trigger_actor, condition_duration, terminate_on_failure=True,
                                         name=condition_name)
@@ -915,19 +761,22 @@ class OpenScenarioParser(object):
 
                     condition_value = ParameterRef(headtime_condition.attrib.get('value'))
 
-                    condition_rule = headtime_condition.attrib.get('rule')
+                    condition_rule = str(ParameterRef(headtime_condition.attrib.get('rule')))
                     condition_operator = OpenScenarioParser.operators[condition_rule]
 
-                    condition_freespace = strtobool(headtime_condition.attrib.get('freespace', False))
-                    condition_along_route = strtobool(headtime_condition.attrib.get('alongRoute', False))
+                    condition_freespace = strtobool(
+                        str(ParameterRef(headtime_condition.attrib.get('freespace', False))))
+                    condition_along_route = strtobool(
+                        str(ParameterRef(headtime_condition.attrib.get('alongRoute', False))))
 
                     for actor in actor_list:
-                        if headtime_condition.attrib.get('entityRef', None) == actor.attributes['role_name']:
+                        if (str(ParameterRef(headtime_condition.attrib.get('entityRef', None)))
+                                == actor.attributes['role_name']):
                             triggered_actor = actor
                             break
                     if triggered_actor is None:
                         raise AttributeError("Cannot find actor '{}' for condition".format(
-                            headtime_condition.attrib.get('entityRef', None)))
+                            str(ParameterRef(headtime_condition.attrib.get('entityRef', None)))))
 
                     atomic = InTimeToArrivalToVehicle(
                         trigger_actor, triggered_actor, condition_value, condition_freespace,
@@ -937,15 +786,15 @@ class OpenScenarioParser(object):
                 elif entity_condition.find('TimeToCollisionCondition') is not None:
                     ttc_condition = entity_condition.find('TimeToCollisionCondition')
 
-                    condition_rule = ttc_condition.attrib.get('rule')
+                    condition_rule = str(ParameterRef(ttc_condition.attrib.get('rule')))
                     condition_operator = OpenScenarioParser.operators[condition_rule]
 
                     condition_value = ParameterRef(ttc_condition.attrib.get('value'))
                     condition_target = ttc_condition.find('TimeToCollisionConditionTarget')
                     entity_ref_ = condition_target.find('EntityRef')
 
-                    condition_freespace = strtobool(ttc_condition.attrib.get('freespace', False))
-                    condition_along_route = strtobool(ttc_condition.attrib.get('alongRoute', False))
+                    condition_freespace = strtobool(str(ParameterRef(ttc_condition.attrib.get('freespace', False))))
+                    condition_along_route = strtobool(str(ParameterRef(ttc_condition.attrib.get('alongRoute', False))))
 
                     if condition_target.find('Position') is not None:
                         position = condition_target.find('Position')
@@ -953,12 +802,13 @@ class OpenScenarioParser(object):
                             trigger_actor, position, condition_value, condition_along_route, condition_operator)
                     else:
                         for actor in actor_list:
-                            if entity_ref_.attrib.get('entityRef', None) == actor.attributes['role_name']:
+                            if (str(ParameterRef(entity_ref_.attrib.get('entityRef', None)))
+                                    == actor.attributes['role_name']):
                                 triggered_actor = actor
                                 break
                         if triggered_actor is None:
                             raise AttributeError("Cannot find actor '{}' for condition".format(
-                                entity_ref_.attrib.get('entityRef', None)))
+                                str(ParameterRef(entity_ref_.attrib.get('entityRef', None)))))
 
                         atomic = InTimeToArrivalToVehicle(
                             trigger_actor, triggered_actor, condition_value, condition_freespace,
@@ -966,18 +816,18 @@ class OpenScenarioParser(object):
                 elif entity_condition.find('AccelerationCondition') is not None:
                     accel_condition = entity_condition.find('AccelerationCondition')
                     condition_value = ParameterRef(accel_condition.attrib.get('value'))
-                    condition_rule = accel_condition.attrib.get('rule')
+                    condition_rule = str(ParameterRef(accel_condition.attrib.get('rule')))
                     condition_operator = OpenScenarioParser.operators[condition_rule]
                     atomic = TriggerAcceleration(
                         trigger_actor, condition_value, condition_operator, condition_name)
                 elif entity_condition.find('StandStillCondition') is not None:
                     ss_condition = entity_condition.find('StandStillCondition')
-                    duration = ParameterRef(ss_condition.attrib.get('duration'))
+                    duration = float(ParameterRef(ss_condition.attrib.get('duration')))
                     atomic = StandStill(trigger_actor, condition_name, duration)
                 elif entity_condition.find('SpeedCondition') is not None:
                     spd_condition = entity_condition.find('SpeedCondition')
                     condition_value = ParameterRef(spd_condition.attrib.get('value'))
-                    condition_rule = spd_condition.attrib.get('rule')
+                    condition_rule = str(ParameterRef(spd_condition.attrib.get('rule')))
                     condition_operator = OpenScenarioParser.operators[condition_rule]
 
                     atomic = TriggerVelocity(
@@ -985,43 +835,45 @@ class OpenScenarioParser(object):
                 elif entity_condition.find('RelativeSpeedCondition') is not None:
                     relspd_condition = entity_condition.find('RelativeSpeedCondition')
                     condition_value = ParameterRef(relspd_condition.attrib.get('value'))
-                    condition_rule = relspd_condition.attrib.get('rule')
+                    condition_rule = str(ParameterRef(relspd_condition.attrib.get('rule')))
                     condition_operator = OpenScenarioParser.operators[condition_rule]
 
                     for actor in actor_list:
-                        if relspd_condition.attrib.get('entityRef', None) == actor.attributes['role_name']:
+                        if (str(ParameterRef(relspd_condition.attrib.get('entityRef', None)))
+                                == actor.attributes['role_name']):
                             triggered_actor = actor
                             break
 
                     if triggered_actor is None:
                         raise AttributeError("Cannot find actor '{}' for condition".format(
-                            relspd_condition.attrib.get('entityRef', None)))
+                            str(ParameterRef(relspd_condition.attrib.get('entityRef', None)))))
 
                     atomic = RelativeVelocityToOtherActor(
                         trigger_actor, triggered_actor, condition_value, condition_operator, condition_name)
                 elif entity_condition.find('TraveledDistanceCondition') is not None:
                     distance_condition = entity_condition.find('TraveledDistanceCondition')
-                    distance_value = ParameterRef(distance_condition.attrib.get('value'))
+                    distance_value = float(ParameterRef(distance_condition.attrib.get('value')))
                     atomic = DriveDistance(trigger_actor, distance_value, name=condition_name)
                 elif entity_condition.find('ReachPositionCondition') is not None:
                     rp_condition = entity_condition.find('ReachPositionCondition')
-                    distance_value = ParameterRef(rp_condition.attrib.get('tolerance'))
+                    distance_value = float(ParameterRef(rp_condition.attrib.get('tolerance')))
                     position = rp_condition.find('Position')
                     atomic = InTriggerDistanceToOSCPosition(
                         trigger_actor, position, distance_value, name=condition_name)
                 elif entity_condition.find('DistanceCondition') is not None:
                     distance_condition = entity_condition.find('DistanceCondition')
 
-                    distance_value = ParameterRef(distance_condition.attrib.get('value'))
+                    distance_value = float(ParameterRef(distance_condition.attrib.get('value')))
 
-                    distance_rule = distance_condition.attrib.get('rule')
+                    distance_rule = str(ParameterRef(distance_condition.attrib.get('rule')))
                     distance_operator = OpenScenarioParser.operators[distance_rule]
 
-                    distance_freespace = strtobool(distance_condition.attrib.get('freespace', False))
+                    distance_freespace = strtobool(str(ParameterRef(distance_condition.attrib.get('freespace', False))))
                     if distance_freespace:
                         raise NotImplementedError(
                             "DistanceCondition: freespace attribute is currently not implemented")
-                    distance_along_route = strtobool(distance_condition.attrib.get('alongRoute', False))
+                    distance_along_route = strtobool(
+                        str(ParameterRef(distance_condition.attrib.get('alongRoute', False))))
 
                     if distance_condition.find('Position') is not None:
                         position = distance_condition.find('Position')
@@ -1031,20 +883,22 @@ class OpenScenarioParser(object):
 
                 elif entity_condition.find('RelativeDistanceCondition') is not None:
                     distance_condition = entity_condition.find('RelativeDistanceCondition')
-                    distance_value = ParameterRef(distance_condition.attrib.get('value'))
+                    distance_value = float(ParameterRef(distance_condition.attrib.get('value')))
 
-                    distance_freespace = distance_condition.attrib.get('freespace', "false") == "true"
-                    rel_dis_type = distance_condition.attrib.get('relativeDistanceType')
+                    distance_freespace = str(ParameterRef(
+                        distance_condition.attrib.get('freespace', "false"))) == "true"
+                    rel_dis_type = str(ParameterRef(distance_condition.attrib.get('relativeDistanceType')))
                     for actor in actor_list:
-                        if distance_condition.attrib.get('entityRef', None) == actor.attributes['role_name']:
+                        if (str(ParameterRef(distance_condition.attrib.get('entityRef', None)))
+                                == actor.attributes['role_name']):
                             triggered_actor = actor
                             break
 
                     if triggered_actor is None:
                         raise AttributeError("Cannot find actor '{}' for condition".format(
-                            distance_condition.attrib.get('entityRef', None)))
+                            str(ParameterRef(distance_condition.attrib.get('entityRef', None)))))
 
-                    condition_rule = distance_condition.attrib.get('rule')
+                    condition_rule = str(ParameterRef(distance_condition.attrib.get('rule')))
                     condition_operator = OpenScenarioParser.operators[condition_rule]
                     atomic = InTriggerDistanceToVehicle(triggered_actor, trigger_actor, distance_value,
                                                         condition_operator, distance_type=rel_dis_type,
@@ -1053,9 +907,9 @@ class OpenScenarioParser(object):
             value_condition = condition.find('ByValueCondition')
             if value_condition.find('ParameterCondition') is not None:
                 parameter_condition = value_condition.find('ParameterCondition')
-                arg_name = parameter_condition.attrib.get('parameterRef')
+                arg_name = str(ParameterRef(parameter_condition.attrib.get('parameterRef')))
                 value = ParameterRef(parameter_condition.attrib.get('value'))
-                rule = parameter_condition.attrib.get('rule')
+                rule = str(ParameterRef(parameter_condition.attrib.get('rule')))
                 if condition_name.startswith('criteria_'):
                     if str(value) != '':
                         arg_value = float(value)
@@ -1079,19 +933,19 @@ class OpenScenarioParser(object):
             elif value_condition.find('SimulationTimeCondition') is not None:
                 simtime_condition = value_condition.find('SimulationTimeCondition')
                 value = ParameterRef(simtime_condition.attrib.get('value'))
-                rule = OpenScenarioParser.operators[simtime_condition.attrib.get('rule')]
+                rule = OpenScenarioParser.operators[str(ParameterRef(simtime_condition.attrib.get('rule')))]
                 atomic = SimulationTimeCondition(value, comparison_operator=rule)
             elif value_condition.find('TimeOfDayCondition') is not None:
                 tod_condition = value_condition.find('TimeOfDayCondition')
-                condition_date = tod_condition.attrib.get('dateTime')
-                condition_rule = tod_condition.attrib.get('rule')
+                condition_date = str(ParameterRef(tod_condition.attrib.get('dateTime')))
+                condition_rule = str(ParameterRef(tod_condition.attrib.get('rule')))
                 condition_operator = OpenScenarioParser.operators[condition_rule]
                 atomic = TimeOfDayComparison(condition_date, condition_operator, condition_name)
             elif value_condition.find('StoryboardElementStateCondition') is not None:
                 state_condition = value_condition.find('StoryboardElementStateCondition')
-                element_name = state_condition.attrib.get('storyboardElementRef')
-                element_type = state_condition.attrib.get('storyboardElementType')
-                state = state_condition.attrib.get('state')
+                element_name = str(ParameterRef(state_condition.attrib.get('storyboardElementRef')))
+                element_type = str(ParameterRef(state_condition.attrib.get('storyboardElementType')))
+                state = str(ParameterRef(state_condition.attrib.get('state')))
                 if state == "startTransition":
                     atomic = OSCStartEndCondition(element_type, element_name, rule="START", name=state + "Condition")
                 elif state in ["stopTransition", "endTransition", "completeState"]:
@@ -1104,10 +958,10 @@ class OpenScenarioParser(object):
             elif value_condition.find('TrafficSignalCondition') is not None:
                 tl_condition = value_condition.find('TrafficSignalCondition')
 
-                name_condition = tl_condition.attrib.get('name')
+                name_condition = str(ParameterRef(tl_condition.attrib.get('name')))
                 traffic_light = OpenScenarioParser.get_traffic_light_from_osc_name(name_condition)
 
-                tl_state = tl_condition.attrib.get('state').upper()
+                tl_state = str(ParameterRef(tl_condition.attrib.get('state'))).upper()
                 if tl_state not in OpenScenarioParser.tl_states:
                     raise KeyError("CARLA only supports Green, Red, Yellow or Off")
                 state_condition = OpenScenarioParser.tl_states[tl_state]
@@ -1135,7 +989,6 @@ class OpenScenarioParser(object):
     def convert_maneuver_to_atomic(action, actor, actor_list, catalogs):
         """
         Convert an OpenSCENARIO maneuver action into a Behavior atomic
-
         Note not all OpenSCENARIO actions are currently supported
         """
         maneuver_name = str(ParameterRef(action.attrib.get('name', 'unknown')))
@@ -1147,10 +1000,10 @@ class OpenScenarioParser(object):
                 if infrastructure_action.find('TrafficSignalStateAction') is not None:
                     traffic_light_action = infrastructure_action.find('TrafficSignalStateAction')
 
-                    name_condition = traffic_light_action.attrib.get('name')
+                    name_condition = str(ParameterRef(traffic_light_action.attrib.get('name')))
                     traffic_light = OpenScenarioParser.get_traffic_light_from_osc_name(name_condition)
 
-                    tl_state = traffic_light_action.attrib.get('state').upper()
+                    tl_state = str(ParameterRef(traffic_light_action.attrib.get('state'))).upper()
                     if tl_state not in OpenScenarioParser.tl_states:
                         raise KeyError("CARLA only supports Green, Red, Yellow or Off")
                     traffic_light_state = OpenScenarioParser.tl_states[tl_state]
@@ -1191,8 +1044,9 @@ class OpenScenarioParser(object):
         elif action.find('UserDefinedAction') is not None:
             user_defined_action = action.find('UserDefinedAction')
             if user_defined_action.find('CustomCommandAction') is not None:
-                command = str(ParameterRef(user_defined_action.find('CustomCommandAction').attrib.get('type')))
+                command = user_defined_action.find('CustomCommandAction').attrib.get('type')
                 atomic = RunScript(command, base_path=OpenScenarioParser.osc_filepath, name=maneuver_name)
+
         elif action.find('PrivateAction') is not None:
             private_action = action.find('PrivateAction')
 
@@ -1205,7 +1059,8 @@ class OpenScenarioParser(object):
                     # duration and distance
                     distance = float('inf')
                     duration = float('inf')
-                    dimension = long_maneuver.find("SpeedActionDynamics").attrib.get('dynamicsDimension')
+                    dimension = str(ParameterRef(long_maneuver.find(
+                        "SpeedActionDynamics").attrib.get('dynamicsDimension')))
                     if dimension == "distance":
                         distance = ParameterRef(long_maneuver.find("SpeedActionDynamics").attrib.get(
                             'value', float("inf")))
@@ -1223,10 +1078,10 @@ class OpenScenarioParser(object):
                     # relative velocity to given actor
                     if long_maneuver.find("SpeedActionTarget").find("RelativeTargetSpeed") is not None:
                         relative_speed = long_maneuver.find("SpeedActionTarget").find("RelativeTargetSpeed")
-                        obj = relative_speed.attrib.get('entityRef')
+                        obj = str(ParameterRef(relative_speed.attrib.get('entityRef')))
                         value = ParameterRef(relative_speed.attrib.get('value', 0))
-                        value_type = relative_speed.attrib.get('speedTargetValueType')
-                        continuous = bool(strtobool(relative_speed.attrib.get('continuous')))
+                        value_type = str(ParameterRef(relative_speed.attrib.get('speedTargetValueType')))
+                        continuous = bool(strtobool(str(ParameterRef(relative_speed.attrib.get('continuous')))))
 
                         for traffic_actor in actor_list:
                             if (traffic_actor is not None and 'role_name' in traffic_actor.attributes and
@@ -1245,7 +1100,7 @@ class OpenScenarioParser(object):
 
                 elif private_action.find('LongitudinalDistanceAction') is not None:
                     long_dist_action = private_action.find("LongitudinalDistanceAction")
-                    obj = long_dist_action.attrib.get('entityRef')
+                    obj = str(ParameterRef(long_dist_action.attrib.get('entityRef')))
                     for traffic_actor in actor_list:
                         if (traffic_actor is not None and 'role_name' in traffic_actor.attributes and
                                 traffic_actor.attributes['role_name'] == obj):
@@ -1260,9 +1115,10 @@ class OpenScenarioParser(object):
                                          "Please specify any one attribute of [distance, timeGap]")
 
                     constraints = long_dist_action.find('DynamicConstraints')
-                    max_speed = constraints.attrib.get('maxSpeed', None) if constraints is not None else None
-                    continues = bool(strtobool(long_dist_action.attrib.get('continuous')))
-                    freespace = bool(strtobool(long_dist_action.attrib.get('freespace')))
+                    max_speed = float(ParameterRef(constraints.attrib.get('maxSpeed', None))
+                                      ) if constraints is not None else None
+                    continues = bool(strtobool(str(ParameterRef(long_dist_action.attrib.get('continuous')))))
+                    freespace = bool(strtobool(str(ParameterRef(long_dist_action.attrib.get('freespace')))))
                     atomic = KeepLongitudinalGap(actor, reference_actor=obj_actor, gap=gap, gap_type=gap_type,
                                                  max_speed=max_speed, continues=continues, freespace=freespace,
                                                  name=maneuver_name)
@@ -1280,7 +1136,8 @@ class OpenScenarioParser(object):
                     # duration and distance
                     distance = float('inf')
                     duration = float('inf')
-                    dimension = lat_maneuver.find("LaneChangeActionDynamics").attrib.get('dynamicsDimension')
+                    dimension = str(ParameterRef(lat_maneuver.find(
+                        "LaneChangeActionDynamics").attrib.get('dynamicsDimension')))
                     if dimension == "distance":
                         distance = ParameterRef(
                             lat_maneuver.find("LaneChangeActionDynamics").attrib.get('value', float("inf")))
@@ -1294,7 +1151,7 @@ class OpenScenarioParser(object):
                                                       name=maneuver_name)
                 elif private_action.find('LaneOffsetAction') is not None:
                     lat_maneuver = private_action.find('LaneOffsetAction')
-                    continuous = bool(strtobool(lat_maneuver.attrib.get('continuous', "true")))
+                    continuous = bool(strtobool(str(ParameterRef(lat_maneuver.attrib.get('continuous', "true")))))
                     # Parsing of the different Dynamic shapes is missing
 
                     lane_target_offset = lat_maneuver.find('LaneOffsetTarget')
@@ -1309,7 +1166,7 @@ class OpenScenarioParser(object):
                         relative_offset = ParameterRef(relative_target_offset.attrib.get('value', 0))
 
                         relative_actor = None
-                        relative_actor_name = relative_target_offset.attrib.get('entityRef', None)
+                        relative_actor_name = str(ParameterRef(relative_target_offset.attrib.get('entityRef', None)))
                         for _actor in actor_list:
                             if _actor is not None and 'role_name' in _actor.attributes:
                                 if relative_actor_name == _actor.attributes['role_name']:
@@ -1333,13 +1190,14 @@ class OpenScenarioParser(object):
                 master_actor = None
                 for actor_ins in actor_list:
                     if actor_ins is not None and 'role_name' in actor_ins.attributes:
-                        if sync_action.attrib.get('masterEntityRef', None) == actor_ins.attributes['role_name']:
+                        if (str(ParameterRef(sync_action.attrib.get('masterEntityRef', None)))
+                                == actor_ins.attributes['role_name']):
                             master_actor = actor_ins
                             break
 
                 if master_actor is None:
                     raise AttributeError("Cannot find actor '{}' for condition".format(
-                        sync_action.attrib.get('masterEntityRef', None)))
+                        str(ParameterRef(sync_action.attrib.get('masterEntityRef', None)))))
 
                 master_position = OpenScenarioParser.convert_position_to_transform(
                     sync_action.find('TargetPositionMaster'))
@@ -1355,7 +1213,7 @@ class OpenScenarioParser(object):
                 elif sync_action.find("FinalSpeed").find("RelativeSpeedToMaster") is not None:
                     relative_speed = sync_action.find("FinalSpeed").find("RelativeSpeedToMaster")
                     final_speed = ParameterRef(relative_speed.attrib.get('value', 0))
-                    relative_type = relative_speed.attrib.get('speedTargetValueType')
+                    relative_type = float(ParameterRef(relative_speed.attrib.get('speedTargetValueType')))
                     atomic = SyncArrivalOSC(
                         actor, master_actor, position, master_position, final_speed,
                         relative_to_master=True, relative_type=relative_type, name=maneuver_name)
@@ -1363,7 +1221,7 @@ class OpenScenarioParser(object):
                     raise AttributeError("Unknown speed action")
             elif private_action.find('ActivateControllerAction') is not None:
                 private_action = private_action.find('ActivateControllerAction')
-                activate = strtobool(private_action.attrib.get('longitudinal'))
+                activate = strtobool(str(ParameterRef(private_action.attrib.get('longitudinal'))))
                 atomic = ChangeAutoPilot(actor, activate, name=maneuver_name)
             elif private_action.find('ControllerAction') is not None:
                 controller_action = private_action.find('ControllerAction')
